deleted unnecessary files
